FailFast: Remove the decored new log functions on the top level logging package
This was unnecessary because we would never expect someone to write
`logging.debug3('foobar')` and the way `logging.configure` works, it will
still adjust the root handler created by `logging.basicConfig`. This was
also causing problems with the short-circuit above because there is no
isEnabledFor function on the top-level logging module object.