PyFilenameFormat: Implement support for skipping alog in stack frames in 3.6+
This is a bit gnarly as the stacklevel argument was introduced in 3.8 and
is a clean/easy way to solve this. Since we support 3.6 and 3.7 too, we
need a "universal" solution. The solution here is based on manual
inspection of the source code of the logging library. While that's
generally a bad idea, it only applies to older versions where that code
will never change.

https://github.com/python/cpython/blob/v3.6.15/Lib/logging/__init__.py#L180
https://github.com/python/cpython/blob/3.7/Lib/logging/__init__.py#L169

https://github.com/IBM/alchemy-logging/issues/152